v2/requirements.txt
# Blocket Bot v2 - Requirements
# Core dependencies
streamlit>=1.28.0
blocket-api>=0.4.3
tenacity>=8.0.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0

# Database
mysql-connector-python>=8.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
isort>=5.12.0
